Introduce subcollection_arbitrary_resource_path :grimace:
This allows "arbitrary resource paths" on subcollections, where
previously they only worked with collections. The reason to add this is
because you may, for instance, wish to return something like a
resource's settings, like this:

```
GET /resource/:id/settings
```

Arbitrary resource paths are considered paths that don't follow the
conventions set forth in the api.yaml with collections and
subcollections. It's very important to note that they actually follow
more standard ActionDispatch conventions in that it's really just a path
routing to a controller with an #index/#show method.

**To be clear: I actually think this is a terrible idea.**
It's mostly to demonstrate the rigidity that the api.yml imposes on
routing, and how deep the rabbit hole is getting with our branding of
what would be very simple routing with raw ActionDispatch methods in to
these confusing labels we place on them (`subcollection_arbitrary_resource_path`? ugh)